refactor: slice 패턴 적용
